initiate timeline switch to a canvas
